prima versione eseguibile

+++ src/main/java/com/treno/application/filter/TrenoFilter.java
@@ -2,16 +2,16 @@ package com.treno.application.filter;
 
 public class TrenoFilter {
 
-	private double prezzoMin, prezzoMax;
-	private double pesoMin, pesoMax;
-	private double lunghezzaMin, lunghezzaMax;
+	private Integer prezzoMin, prezzoMax;
+	private Integer pesoMin, pesoMax;
+	private Integer lunghezzaMin, lunghezzaMax;
 	private String sigla;
 	
 	public double getPrezzoMin() {
 		return prezzoMin;
 	}
 	
-	public void setPrezzoMin(double prezzoMin) {
+	public void setPrezzoMin(int prezzoMin) {
 		this.prezzoMin = prezzoMin;
 	}
 	
@@ -19,7 +19,7 @@ public class TrenoFilter {
 		return prezzoMax;
 	}
 	
-	public void setPrezzoMax(double prezzoMax) {
+	public void setPrezzoMax(int prezzoMax) {
 		this.prezzoMax = prezzoMax;
 	}
 	
@@ -27,7 +27,7 @@ public class TrenoFilter {
 		return pesoMin;
 	}
 	
-	public void setPesoMin(double pesoMin) {
+	public void setPesoMin(int pesoMin) {
 		this.pesoMin = pesoMin;
 	}
 	
@@ -35,7 +35,7 @@ public class TrenoFilter {
 		return pesoMax;
 	}
 	
-	public void setPesoMax(double pesoMax) {
+	public void setPesoMax(int pesoMax) {
 		this.pesoMax = pesoMax;
 	}
 	
@@ -43,7 +43,7 @@ public class TrenoFilter {
 		return lunghezzaMin;
 	}
 	
-	public void setLunghezzaMin(double lunghezzaMin) {
+	public void setLunghezzaMin(int lunghezzaMin) {
 		this.lunghezzaMin = lunghezzaMin;
 	}
 	
@@ -51,7 +51,7 @@ public class TrenoFilter {
 		return lunghezzaMax;
 	}
 	
-	public void setLunghezzaMax(double lunghezzaMax) {
+	public void setLunghezzaMax(int lunghezzaMax) {
 		this.lunghezzaMax = lunghezzaMax;
 	}
